incident-iq-mvp-v1/requirements.txt
# Python 3.11
streamlit==1.38.0
azure-identity==1.17.1
azure-search-documents
azure-ai-contentsafety==1.0.0
azure-core==1.30.2
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.1
tiktoken==0.7.0
tenacity==8.5.0
httpx==0.27.0